requirements_gpu_full.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cuda-bindings==12.9.4
cuda-core==0.3.2
cuda-pathfinder==1.3.2
cuda-python==12.9.4
cuda-toolkit==12.9.1
cudf-cu12==24.4.1
cuml-cu12==24.4.0
cupy-cuda12x==13.6.0
cuvs-cu12==24.10.0
dask==2024.1.1
dask-cuda==24.4.0
dask-cudf-cu12==24.4.1
dask-expr==0.4.0
distributed==2024.1.1
distributed-ucxx-cu12==0.40.0
fastrlock==0.8.3
flatbuffers==25.9.23
fsspec==2025.10.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
libcudf-cu12==24.10.1
libcuml-cu12==25.10.0
libkvikio-cu12==25.10.0
libraft-cu12==25.10.0
librmm-cu12==25.10.0
libucx-cu12==1.17.0.post1
libucxx-cu12==0.40.0
lightgbm==4.6.0
llvmlite==0.45.1
locket==1.0.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
namex==0.1.0
numba==0.62.1
numba-cuda==0.19.1
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.16.0.29
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.9.86
nvtx==0.2.14
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.2.1
partd==1.4.2
pathspec==0.12.1
pillow==12.0.0
protobuf==6.33.1
psutil==7.1.3
pyarrow==14.0.2
Pygments==2.19.2
pylibcudf-cu12==24.10.1
pylibraft-cu12==24.4.0
pynvjitlink-cu12==0.7.0
pynvml==11.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
raft-dask-cu12==24.4.0
rapids-dask-dependency==24.4.1
rapids-logger==0.1.19
requests==2.32.5
rich==14.2.0
rmm-cu12==24.4.0
scikit-learn==1.7.2
scikit_build_core==0.11.6
scipy==1.16.3
six==1.17.0
sortedcontainers==2.4.0
tblib==3.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.5.0
toolz==1.1.0
tornado==6.5.2
treelite==4.1.2
typing_extensions==4.15.0
tzdata==2025.2
ucx-py-cu12==0.37.0
ucxx-cu12==0.40.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==2.0.1
xgboost==3.1.2
zict==3.0.0
zipp==3.23.0
